Add analytic solution to the problem
Part 2 of the problem looked like it was going
to stress the brute force, check all values,
solution. So, before tackling Part 2, i've
implemented an analytic solutions that should
be constant time rather than O(n). It works
for Part 1. Just have to implement the parsing
for Part 2 so I can check it works there.